Extract `create_market` tests
